+++ lib/mix/test/mix/cli_test.exs
@@ -3,20 +3,38 @@ Code.require_file "../test_helper.exs", __DIR__
 defmodule Mix.CLITest do
   use MixTest.Case
 
-  test "env" do
-    in_fixture "only_mixfile", fn ->
-      if match? {:win32, _}, :os.type do
-        temp_env = "set MIX_ENV=prod &"
-      else
-        temp_env = "MIX_ENV=prod"
+  test "env configs" do
+    in_fixture "no_mixfile", fn ->
+      File.write! "custom.exs", """
+      defmodule P do
+        use Mix.Project
+        def project, do: [app: :p, version: "0.1.0"]
       end
+      """
 
-      env = System.cmd ~s(#{temp_env} #{elixir_executable} #{mix_executable} run -e "IO.inspect {Mix.env, System.argv}" -- 1 2 3)
-      assert env =~ ~s({:prod, ["1", "2", "3"]})
+      output = System.cmd ~s(MIX_ENV=prod MIX_EXS=custom.exs #{elixir_executable} #{mix_executable}) <>
+                          ~s( run -e "IO.inspect {Mix.env, System.argv}" -- 1 2 3)
+
+      assert output =~ ~s({:prod, ["1", "2", "3"]})
+      assert output =~ "Compiled lib/a.ex"
     end
   end
 
   test "default task" do
+    in_fixture "no_mixfile", fn ->
+      File.write! "mix.exs", """
+      defmodule P do
+        use Mix.Project
+        def project, do: [app: :p, version: "0.1.0"]
+      end
+      """
+      output = mix ""
+      assert File.regular?("_build/dev/lib/p/ebin/Elixir.A.beam")
+      assert output =~ "Compiled lib/a.ex"
+    end
+  end
+
+  test "custom mix.exs" do
     in_fixture "no_mixfile", fn ->
       File.write! "mix.exs", """
       defmodule P do